requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dbf-light==1.0.0
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
gTTS==2.5.1
html5lib==1.1
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
magic-filter==1.0.12
matplotlib==3.9.1
multidict==6.0.5
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.10.2
pycbrf==1.1.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
pyshorteners==1.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.40
